feat(review-improvement-loop): コードベース総合レビュー＆改善ループ完走（64 セル）
移植可能・再実行型のコード総合レビュー＆改善ループを全 64 セル完走。
領域自己発見 × 7 次元（G1 テスト網羅／G2 静的衛生／G3 ハードニング／
G4 簡素化／G5 文書・依存整合）マトリクスで全資産を点検・改善し、
各セルを独立レビューでゲートした。外部仕様（DSL 文法・Lua API・
SHIORI 正常系）は不変更で、挙動変化は不正入力経路のハードニング 19 件
（全件境界回帰テスト付き）のみに限定。

主な成果:
- Rust クレート: pasta_core/pasta_dsl/pasta_lua/pasta_shiori/pasta_lsp/
  pasta_check/pasta_sample_ghost へテスト大量追加（Rust 1510→1956・+446）・
  clippy 全ゼロ化・到達可能パニック経路の Result 化
- ハードニング 19 件: FFI/SHIORI 境界の脆弱性 6 件・frameId オーバーフロー・
  Content-Length DoS ガード・loader トラバーサル・ZIP パストラバーサル検査の
  release 常時有効化・逆転 range パニック硬化・Lua try_route nil ガード・
  book リンク検証のパストラバーサルガード・bigram-index Windows CLI 修正 等
- Lua 資産: lua_specs へ 96 テスト追加（44→45 スイート）・luacheck 導入で
  警告ゼロ化・pasta_scripts/shiori 層の sweep ヘッダー正準化
- VSCode 拡張: 休眠モック初配線による実モジュールテスト基盤新設・
  eslint 導入（37→0）・デバッグポート/wasm 境界型ガード・ドリフト済み
  コピー型テスト整理（実カバレッジ純増）
- book/tools: drift/tutorial-check・highlight・bigram-index へテスト網羅
- D7 横断: cargo audit（imageproc unsound 3 件解消）・cargo-deny/machete 導入・
  npm audit（vscode 15→1）・TEST_COVERAGE.md 全面実測同期（台帳 2091）
- 改善レポート生成（reports/2026-06-12-improvement-report.md・必須 5 セクション）

導入ツール（dev 環境のみ・リポジトリ非変更）: luacheck 1.2.0・
cargo-deny 0.19.8・cargo-machete 0.9.2。リポジトリ変更: eslint（vscode devDeps）。

本仕様は再実行型プロセス仕様につき completed/ へ移動せず .kiro/specs/ 直下に留まる。

+++ crates/pasta_lua/tests/transpiler/record_wiring_test.rs
@@ -54,7 +54,7 @@ impl SourceMapSink for LinePairSink {
 /// 既知 `.pasta` 行を持つ有効 span（end_byte > 0）を作る小ヘルパ。
 fn span_at(line: usize) -> Span {
     // start_line/col, end_line/col, start_byte, end_byte（end_byte>0 で is_valid）
-    Span::new(line, 1, line, 10, (line * 10) as usize, (line * 10 + 9) as usize)
+    Span::new(line, 1, line, 10, line * 10, line * 10 + 9)
 }
 
 // ============================================================================
